+++ app-backend/batch/src/main/scala/stacExport/LabelCollectionBuilder.scala
@@ -136,10 +136,13 @@ class LabelCollectionBuilder[
       tasks: List[Task],
       tasksGeomExtent: Option[UnionedGeomExtent]
   ): LabelCollectionBuilder[
-    CollectionRequirements with CollectionTasksGeomExtent] =
+    CollectionRequirements with CollectionTasksGeomExtent
+  ] =
     new LabelCollectionBuilder(
-      labelCollection.copy(tasks = labelCollection.tasks ++ tasks,
-                           tasksGeomExtent = tasksGeomExtent)
+      labelCollection.copy(
+        tasks = labelCollection.tasks ++ tasks,
+        tasksGeomExtent = tasksGeomExtent
+      )
     )
 
   def withItemPropInfo(
@@ -195,31 +198,34 @@ class LabelCollectionBuilder[
         labelItemSelfAbsLink,
         Self,
         Some(`application/json`),
-        Some(s"Label item ${labelItemId}")
+        Some(s"Label item ${labelItemId}"),
+        List(labelItemId)
       ),
       StacLink(
         "../collection.json",
         Parent,
         Some(`application/json`),
-        Some("Label Collection")
+        Some("Label Collection"),
+        List()
       ),
       StacLink(
         s"../${rootPath}",
         StacRoot,
         Some(`application/json`),
-        Some("Root")
+        Some("Root"),
+        List()
       )
     ) ++ labelCollection.sceneItemLinks.map(link => {
-      // TODO: For the rel (the second argumetn), we need a Source type,
-      // which needs to be added in gt-server
       StacLink(
         s"../${link._2}",
-        VendorLinkType("Source"),
+        Source,
         Some(`image/cog`),
-        Some("Source image STAC item for the label item")
+        Some("Source image STAC item for the label item"),
+        List()
       )
     })
-    val dateTime = labelCollection.extent.get.temporal(0) match {
+    val dateTime
+      : List[Timestamp] = labelCollection.extent.get.temporal.interval.flatten map {
       case Some(dt) => Timestamp.from(Instant.parse(s"${dt}Z"))
       case _        => new Timestamp(new java.util.Date().getTime)
     }
@@ -231,7 +237,7 @@ class LabelCollectionBuilder[
       Some(List(labelCollection.itemPropsThin.task)),
       Some(List("manual")),
       None,
-      dateTime
+      dateTime(0)
     )
     val labelItemPropertiesJsonObj = JsonObject.fromMap(
       Map(
@@ -245,9 +251,6 @@ class LabelCollectionBuilder[
         "datetime" -> labelItemProperties.datetime.asJson
       )
     )
-    // TODO: below should actually be `application/geo+json`
-    // but StacItem from gt-server only accepts `image/cog`
-    // or it will throw an exception
     val labelDataRelLink = "./data.geojson"
     val labelDataS3AbsLink: String = s"${labelItemSelfAbsPath}/data.geojson"
     val labelAsset = Map(
@@ -255,9 +258,10 @@ class LabelCollectionBuilder[
         StacAsset(
           labelDataRelLink,
           Some("Label Data Feature Collection"),
-          Some(`image/cog`)
+          Some(`application/geo+json`)
         )
     )
+    val stacExtensions = List("label")
     val labelItem: StacItem = itemBuilder
       .withId(labelItemId)
       .withGeometries(labelItemFootprint, labelItemBbox)
@@ -266,6 +270,8 @@ class LabelCollectionBuilder[
       .withProperties(labelItemPropertiesJsonObj)
       .withParentPath(absPath, rootPath)
       .withAssets(labelAsset)
+      .withStacVersion(labelCollection.stacVersion)
+      .withExtensions(stacExtensions)
       .build()
 
     (
@@ -276,7 +282,8 @@ class LabelCollectionBuilder[
               labelItemSelfRelLink,
               Item,
               Some(`application/json`),
-              Some("STAC label item link")
+              Some("STAC label item link"),
+              List()
             )
           )
         )